+++ packages/block-library/src/index.js
@@ -20,6 +20,7 @@ import * as gallery from './gallery';
 import * as archives from './archives';
 import * as audio from './audio';
 import * as button from './button';
+import * as calendar from './calendar';
 import * as categories from './categories';
 import * as code from './code';
 import * as columns from './columns';
@@ -38,6 +39,8 @@ import * as nextpage from './nextpage';
 import * as preformatted from './preformatted';
 import * as pullquote from './pullquote';
 import * as reusableBlock from './block';
+import * as rss from './rss';
+import * as search from './search';
 import * as separator from './separator';
 import * as shortcode from './shortcode';
 import * as spacer from './spacer';
@@ -67,6 +70,7 @@ export const registerCoreBlocks = () => {
 		archives,
 		audio,
 		button,
+		calendar,
 		categories,
 		code,
 		columns,
@@ -86,6 +90,8 @@ export const registerCoreBlocks = () => {
 		nextpage,
 		preformatted,
 		pullquote,
+		rss,
+		search,
 		separator,
 		reusableBlock,
 		spacer,