+++ packages/block-library/src/table/edit.js
@@ -436,7 +436,7 @@ export class TableEdit extends Component {
 					<Toolbar>
 						<DropdownMenu
 							icon="editor-table"
-							label={ __( 'Edit Table' ) }
+							label={ __( 'Edit table' ) }
 							controls={ this.getTableControls() }
 						/>
 					</Toolbar>