+++ packages/components/src/toolbar/toolbar-container.native.js
@@ -3,10 +3,13 @@
  */
 import { View } from 'react-native';
 
+/**
+ * Internal dependencies
+ */
 import styles from './style.scss';
 
 const ToolbarContainer = ( props ) => (
-	<View style={ styles.container }>
+	<View style={ [ styles.container, props.passedStyle ] }>
 		{ props.children }
 	</View>
 );