+++ packages/eslint-plugin/configs/custom.js
@@ -1,5 +1,11 @@
 module.exports = {
+	plugins: [
+		'@wordpress',
+	],
 	rules: {
+		'@wordpress/dependency-group': 'error',
+		'@wordpress/no-unused-vars-before-return': 'error',
+		'@wordpress/valid-sprintf': 'error',
 		'no-restricted-syntax': [
 			'error',
 			{