+++ packages/redux-routine/src/runtime.js
@@ -8,7 +8,6 @@ import isPromise from 'is-promise';
 /**
  * Internal dependencies
  */
-import castError from './cast-error';
 import { isActionOfType, isAction } from './is-action';
 
 /**
@@ -27,12 +26,9 @@ export default function createRuntime( controls = {}, dispatch ) {
 		const routine = control( value );
 		if ( isPromise( routine ) ) {
 			// Async control routine awaits resolution.
-			routine.then(
-				yieldNext,
-				( error ) => yieldError( castError( error ) ),
-			);
+			routine.then( yieldNext, yieldError );
 		} else {
-			next( routine );
+			yieldNext( routine );
 		}
 		return true;
 	} );